requirements.txt
transformers
torch
python-dotenv
huggingface_hub
